requirements.txt
numpy==1.23.4
pandas==1.5.3
torch==2.0.1
opencv-python==4.7.0.72
Pillow==9.3.0
streamlit==1.19.0
matplotlib==3.6.2
scikit-learn==1.2.2
pyyaml==6.0
tqdm==4.64.1
requests==2.28.1
torchvision==0.15.2
tensorflow==2.13.0
transformers==4.30.2
timm==0.9.2
easyocr==1.6.2
pycocotools==2.0.6
